Include AttrJson::NestedAttributes and AttrJson::QueryScopes in all AttrJson::Record
* You no longer need to explictly `include AttrJson::NestedAttributes` or `include AttrJson::QueryScopes` in your record class. But it shouldn't hurt if you do.

* If you for some reason want an AttrJson::Record *without* those modules included, you can use `AttrJson::Record::Base`. One plausible use case for that might be trying to use AttrJson with MySQL -- where AttrJson::QueryScopes definitely won't work!
